+++ cli/var.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"os"
 
-	"go.bbkane.com/envelope/cli/tableprint"
-	"go.bbkane.com/envelope/models"
+	"go.bbkane.com/enventory/cli/tableprint"
+	"go.bbkane.com/enventory/models"
 	"go.bbkane.com/warg/command"
 	"go.bbkane.com/warg/flag"
 	"go.bbkane.com/warg/value/scalar"